Commit #4
Added forward and backwards action waypoints. Current identifier is
"shoot". Robot program can determine when to execute action. Graphically
the action executes at the end of a movement. So for example: doing a
drive-shoot would require 2 clicks. One for the robot start, one for the
drive, which is done with Click+A. The data that is sent says the bot
needs to turn x degrees and then drive for the specified distance, and
then the action "shoot" could be executed...