Adjust angle for starting point.
We NEED to be able to set this first angle to the Navx if we want it to
not correct wildly. A robot starting at 0.4 rads in program and real
life will make the NavX read 0 on a reset, messing everything up. We
will need to make the NavX's base angle that of the first line of the
.BOND file